requirements.txt
numpy
hidapi
pyserial
serial
pygame
sockets